# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Mealswapp API
  version: 0.3.0
  description: Shared gateway source of truth for Phase 03 authentication and gateway contracts.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /
security: []
paths:
...
  /api/v1/auth/csrf-token:
    get:
      operationId: getApiV1AuthCsrfToken
      summary: Obtain a session-bound CSRF synchronizer token
      responses:
        "200":
          description: CSRF token delivery envelope. HttpOnly CSRF cookies are set separately.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CSRFTokenEnvelope"
        "400":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
...
components:
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: mealswapp_access
    refreshCookie:
      type: apiKey
      in: cookie
      name: mealswapp_refresh
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
  responses:
...
    Error:
      description: User-safe gateway error envelope with no client-visible audit snapshot, raw provider payload, or internal diagnostics.
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    # Implements DESIGN-004 JobStatusTracker and DESIGN-010 RateLimiter shared admission contract.
...
  schemas:
    AppError:
      type: object
      additionalProperties: false
      required: [category, code, message, retryable]
      properties:
        category:
          type: string
          enum: [validation, auth, entitlement, security, network, timeout, server, dependency, rate_limit, unknown]
        code:
          type: string
          minLength: 1
          maxLength: 120
        message:
          type: string
          minLength: 1
          maxLength: 500
        retryable:
          type: boolean
        requestId:
          type: string
          minLength: 1
          maxLength: 120
    # Implements DESIGN-009 AdminController audit-safe error boundary.
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [status, requestId, error]
      properties:
        status:
          type: string
          const: error
        requestId:
          type: string
          maxLength: 120
        error:
          $ref: "#/components/schemas/AppError"
...
    CSRFTokenEnvelope:
      type: object
      required: [status, requestId, data]
      properties:
        status:
          type: string
          enum: [ok]
        requestId:
          type: string
        data:
          type: object
          required: [csrfToken]
          properties:
            csrfToken:
              type: string
...